+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Mega Pinsir animations! & Scizor bullet punch animation changes
 - Dynamax level is not hardcoded anymore for the summary screen
 - Fixed msd spawnpools
+- Fixed pedestal block look direction in structures